+++ src/ua/user_identity_token.rs
@@ -4,13 +4,15 @@ use crate::ua;
 pub enum UserIdentityToken {
     Anonymous(ua::AnonymousIdentityToken),
     UserName(ua::UserNameIdentityToken),
+    X509(ua::X509IdentityToken),
 }
 
 impl UserIdentityToken {
     pub(crate) fn to_extension_object(&self) -> ua::ExtensionObject {
         match self {
-            UserIdentityToken::Anonymous(anonymous) => ua::ExtensionObject::new(anonymous),
-            UserIdentityToken::UserName(user_name) => ua::ExtensionObject::new(user_name),
+            Self::Anonymous(anonymous) => ua::ExtensionObject::new(anonymous),
+            Self::UserName(user_name) => ua::ExtensionObject::new(user_name),
+            Self::X509(x509) => ua::ExtensionObject::new(x509),
         }
     }
 }
@@ -24,5 +26,11 @@ impl From<ua::AnonymousIdentityToken> for UserIdentityToken {
 impl From<ua::UserNameIdentityToken> for UserIdentityToken {
     fn from(value: ua::UserNameIdentityToken) -> Self {
         Self::UserName(value)
+    }
+}
+
+impl From<ua::X509IdentityToken> for UserIdentityToken {
+    fn from(value: ua::X509IdentityToken) -> Self {
+        Self::X509(value)
     }
 }